+++ src/model/squares/Brick.java
@@ -4,7 +4,7 @@ import model.Square;
 
 public class Brick extends Square {
 	public Brick() {
-		super(1);
+		super((byte) 1);
 	}
