+++ src/model/squares/Ladder.java
@@ -4,7 +4,7 @@ import model.Square;
 
 public class Ladder extends Square {
 	public Ladder() {
-		super(4);
+		super((byte) 4);
 	}
 
 }
